+++ app/db/helpers/script.py
@@ -1,8 +1,31 @@
+from typing import Optional
+
 from sqlalchemy.ext.asyncio import AsyncConnection
 from sqlalchemy.sql import func, literal_column, select
 
 from ...models.raw import ScriptFileList
-from ...schemas.raw import ScriptSearchResult
+from ...schemas.raw import ScriptEntity, ScriptSearchResult
+from .quest import get_quest_entity
+
+
+async def get_script(conn: AsyncConnection, script_id: str) -> Optional[ScriptEntity]:
+    stmt = select(
+        ScriptFileList.c.scriptFileName,
+        func.octet_length(ScriptFileList.c.rawScript).label("scriptSizeBytes"),
+        ScriptFileList.c.questId,
+    ).where(ScriptFileList.c.scriptFileName == script_id)
+
+    rows = (await conn.execute(stmt)).fetchall()
+
+    if len(rows) == 0:
+        return None
+
+    quest_ids: list[int] = [row.questId for row in rows]
+    quests = await get_quest_entity(conn, quest_ids)
+
+    return ScriptEntity(
+        scriptId=script_id, scriptSizeBytes=rows[0].scriptSizeBytes, quests=quests
+    )
 
 
 async def get_script_search(