+++ app/rayshift/quest.py
@@ -5,7 +5,7 @@ import httpx
 from fastapi import HTTPException
 from sqlalchemy.ext.asyncio import AsyncConnection
 
-from ..config import SecretSettings, Settings
+from ..config import SecretSettings, Settings, logger
 from ..db.helpers.rayshift import get_rayshift_quest_db, insert_rayshift_quest_db
 from ..schemas.common import Region
 from ..schemas.rayshift import (
@@ -88,7 +88,7 @@ def get_multiple_quests(region: Region, query_ids: list[int]) -> dict[int, Quest
     r = httpx.get(f"{QUEST_ENDPOINT}/get", params=params)
     if r.status_code == httpx.codes.TOO_MANY_REQUESTS:  # pragma: no cover
         sleep_time = r.json().get("wait", 10)
-        print(f"Sleeping {sleep_time} seconds for RayShift API")
+        logger.info(f"Sleeping {sleep_time} seconds for RayShift API")
         time.sleep(sleep_time)
         r = httpx.get(f"{QUEST_ENDPOINT}/get", params=params)